// src/data/cosmeticModels.gen.ts
export const COSMETIC_MODELS: Record<string, number> = {
  emblem_bigben: require('../../assets/models3d/emblem_bigben.glb'),
  emblem_christ: require('../../assets/models3d/emblem_christ.glb'),
  emblem_colosseum: require('../../assets/models3d/emblem_colosseum.glb'),
  emblem_compass: require('../../assets/models3d/emblem_compass.glb'),
  emblem_eiffel: require('../../assets/models3d/emblem_eiffel.glb'),
  emblem_fuji: require('../../assets/models3d/emblem_fuji.glb'),
  emblem_goldengate: require('../../assets/models3d/emblem_goldengate.glb'),
  emblem_greatwall: require('../../assets/models3d/emblem_greatwall.glb'),
  emblem_liberty: require('../../assets/models3d/emblem_liberty.glb'),
  emblem_moai: require('../../assets/models3d/emblem_moai.glb'),
  emblem_pisa: require('../../assets/models3d/emblem_pisa.glb'),
  emblem_pyramids: require('../../assets/models3d/emblem_pyramids.glb'),
  emblem_st_laurel: require('../../assets/models3d/emblem_st_laurel.glb'),
  emblem_st_star: require('../../assets/models3d/emblem_st_star.glb'),
  emblem_st_summit: require('../../assets/models3d/emblem_st_summit.glb'),
  emblem_st_worldtree: require('../../assets/models3d/emblem_st_worldtree.glb'),
  emblem_sydney: require('../../assets/models3d/emblem_sydney.glb'),
  emblem_taj: require('../../assets/models3d/emblem_taj.glb'),
  emblem_windmill: require('../../assets/models3d/emblem_windmill.glb'),
  globe_ice_props: require('../../assets/models3d/globe_ice_props.glb'),
  globe_land: require('../../assets/models3d/globe_land.glb'),
  globe_lava_props: require('../../assets/models3d/globe_lava_props.glb'),
  globe_mars_props: require('../../assets/models3d/globe_mars_props.glb'),
  globe_relief_props: require('../../assets/models3d/globe_relief_props.glb'),
  globe_st_crowned_props: require('../../assets/models3d/globe_st_crowned_props.glb'),
  globe_st_fractured_props: require('../../assets/models3d/globe_st_fractured_props.glb'),
  orbit_asteroids: require('../../assets/models3d/orbit_asteroids.glb'),
  orbit_compass: require('../../assets/models3d/orbit_compass.glb'),
  orbit_double: require('../../assets/models3d/orbit_double.glb'),
  orbit_fire: require('../../assets/models3d/orbit_fire.glb'),
  orbit_fireflies: require('../../assets/models3d/orbit_fireflies.glb'),
  orbit_graticule: require('../../assets/models3d/orbit_graticule.glb'),
  orbit_ice: require('../../assets/models3d/orbit_ice.glb'),
  orbit_meridian: require('../../assets/models3d/orbit_meridian.glb'),
  orbit_neon: require('../../assets/models3d/orbit_neon.glb'),
  orbit_rainbow: require('../../assets/models3d/orbit_rainbow.glb'),
  orbit_saturn: require('../../assets/models3d/orbit_saturn.glb'),
  orbit_st_compass: require('../../assets/models3d/orbit_st_compass.glb'),
  orbit_st_laurel: require('../../assets/models3d/orbit_st_laurel.glb'),
  sat_balloon: require('../../assets/models3d/sat_balloon.glb'),
  sat_bird: require('../../assets/models3d/sat_bird.glb'),
  sat_comet: require('../../assets/models3d/sat_comet.glb'),
  sat_iss: require('../../assets/models3d/sat_iss.glb'),
  sat_moon: require('../../assets/models3d/sat_moon.glb'),
  sat_paperplane: require('../../assets/models3d/sat_paperplane.glb'),
  sat_plane: require('../../assets/models3d/sat_plane.glb'),
  sat_rocket: require('../../assets/models3d/sat_rocket.glb'),
  sat_satellite: require('../../assets/models3d/sat_satellite.glb'),
  sat_shootingstar: require('../../assets/models3d/sat_shootingstar.glb'),
  sat_st_comet: require('../../assets/models3d/sat_st_comet.glb'),
  sat_st_moon: require('../../assets/models3d/sat_st_moon.glb'),
  sat_st_ship: require('../../assets/models3d/sat_st_ship.glb'),
  sat_ufo: require('../../assets/models3d/sat_ufo.glb'),
};

export const GLOBE_TEXTURES: Record<string, number> = {
  biolum: require('../../assets/globes/globe_biolum.webp'),
  blueprint: require('../../assets/globes/globe_blueprint.webp'),
  classic: require('../../assets/globes/globe_classic.webp'),
  cyber: require('../../assets/globes/globe_cyber.webp'),
  eclipse: require('../../assets/globes/globe_eclipse.webp'),
  gaia: require('../../assets/globes/globe_gaia.webp'),
  gold: require('../../assets/globes/globe_gold.webp'),
  hologram: require('../../assets/globes/globe_hologram.webp'),
  ice: require('../../assets/globes/globe_ice.webp'),
  lava: require('../../assets/globes/globe_lava.webp'),
  mars: require('../../assets/globes/globe_mars.webp'),
  night: require('../../assets/globes/globe_night.webp'),
  pastel: require('../../assets/globes/globe_pastel.webp'),
  political: require('../../assets/globes/globe_political.webp'),
  relief: require('../../assets/globes/globe_relief.webp'),
  satellite: require('../../assets/globes/globe_satellite.webp'),
  st_crowned: require('../../assets/globes/globe_st_crowned.webp'),
  st_fractured: require('../../assets/globes/globe_st_fractured.webp'),
  st_galaxy: require('../../assets/globes/globe_st_galaxy.webp'),
  vintage: require('../../assets/globes/globe_vintage.webp'),
};